:wrench: chore: Arrumando a definição das pastas, e arquivos
